requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
fastapi
azure-communication-email
stripe
azure-data-tables
python-jose[cryptography]
httpx
azure-identity
requests
beautifulsoup4
azure-storage-blob
google-api-python-client
